+++ TheForceEngine/TFE_RenderShared/lineDraw3d.cpp
@@ -187,7 +187,7 @@ namespace TFE_RenderShared
 		s_curveCount3d++;
 		// bounding box.
 		f32 maxOffset = 0.0f;
-		for (s32 i = 0; i < offsetCount; i++)
+		for (s32 i = 0; i < (s32)offsetCount; i++)
 		{
 			maxOffset = std::max(maxOffset, fabsf(offsets[i]));
 		}
@@ -211,7 +211,7 @@ namespace TFE_RenderShared
 
 		// Offsets
 		Vec4f offsetVec = { 0 };
-		for (s32 i = 0; i < offsetCount && i < 4; i++)
+		for (s32 i = 0; i < (s32)offsetCount && i < 4; i++)
 		{
 			offsetVec.m[i] = offsets[i];
 		}
